kabinizer-front-end/src/utils/createWeeksForMonth.test.ts
import { describe, expect, test } from "vitest";
import { createWeeksForMonth } from "@/utils/createWeeksForMonth.ts";

describe("createWeeksForMonth", () => {
  test("January 2024: 31 days starting on Monday", () => {
    const weeks = createWeeksForMonth(0, 2024); // January 2024
    expect(weeks).toEqual([
      [1, 2, 3, 4, 5, 6, 7],
      [8, 9, 10, 11, 12, 13, 14],
      [15, 16, 17, 18, 19, 20, 21],
      [22, 23, 24, 25, 26, 27, 28],
      [29, 30, 31, null, null, null, null],
    ]);
  });

  test("April 2024: 30 days starting on Tuesday", () => {
    const weeks = createWeeksForMonth(3, 2021); // April 2024
    expect(weeks).toEqual([
      [null, null, null, 1, 2, 3, 4],
      [5, 6, 7, 8, 9, 10, 11],
      [12, 13, 14, 15, 16, 17, 18],
      [19, 20, 21, 22, 23, 24, 25],
      [26, 27, 28, 29, 30, null, null],
    ]);
  });

  test("February 2023: 28 days starting on Wednesday", () => {
    const weeks = createWeeksForMonth(1, 2023); // February 2023
    expect(weeks).toEqual([
      [null, null, 1, 2, 3, 4, 5],
      [6, 7, 8, 9, 10, 11, 12],
      [13, 14, 15, 16, 17, 18, 19],
      [20, 21, 22, 23, 24, 25, 26],
      [27, 28, null, null, null, null, null],
    ]);
  });

  test("February 2024: 29 days starting on Thursday", () => {
    const weeks = createWeeksForMonth(1, 2024); // February 2024
    expect(weeks).toEqual([
      [null, null, null, 1, 2, 3, 4],
      [5, 6, 7, 8, 9, 10, 11],
      [12, 13, 14, 15, 16, 17, 18],
      [19, 20, 21, 22, 23, 24, 25],
      [26, 27, 28, 29, null, null, null],
    ]);
  });

  test("Leap Year February 2020", () => {
    const weeks = createWeeksForMonth(1, 2020); // February 2020
    expect(weeks).toEqual([
      [null, null, null, null, null, 1, 2],
      [3, 4, 5, 6, 7, 8, 9],
      [10, 11, 12, 13, 14, 15, 16],
      [17, 18, 19, 20, 21, 22, 23],
      [24, 25, 26, 27, 28, 29, null],
    ]);
  });

  test("March 2028: 31 days starting on Wednesday", () => {
    const weeks = createWeeksForMonth(2, 2028); // March 2028
    expect(weeks).toEqual([
      [null, null, 1, 2, 3, 4, 5],
      [6, 7, 8, 9, 10, 11, 12],
      [13, 14, 15, 16, 17, 18, 19],
      [20, 21, 22, 23, 24, 25, 26],
      [27, 28, 29, 30, 31, null, null],
    ]);
  });

  test("Only add the last week if it contains any days of the current month", () => {
    let weeks = createWeeksForMonth(11, 2028); // December 2028
    expect(weeks).toEqual([
      [null, null, null, null, 1, 2, 3],
      [4, 5, 6, 7, 8, 9, 10],
      [11, 12, 13, 14, 15, 16, 17],
      [18, 19, 20, 21, 22, 23, 24],
      [25, 26, 27, 28, 29, 30, 31],
    ]);

    weeks = createWeeksForMonth(5, 2024); // June 2024
    expect(weeks).toEqual([
      [null, null, null, null, null, 1, 2],
      [3, 4, 5, 6, 7, 8, 9],
      [10, 11, 12, 13, 14, 15, 16],
      [17, 18, 19, 20, 21, 22, 23],
      [24, 25, 26, 27, 28, 29, 30],
    ]);
  });

  test("All weeks in 2024", () => {
    const jan = createWeeksForMonth(0, 2024); // January 2024
    expect(jan).toEqual([
      [1, 2, 3, 4, 5, 6, 7],
      [8, 9, 10, 11, 12, 13, 14],
      [15, 16, 17, 18, 19, 20, 21],
      [22, 23, 24, 25, 26, 27, 28],
      [29, 30, 31, null, null, null, null],
    ]);

    const feb = createWeeksForMonth(1, 2024); // February 2024
    expect(feb).toEqual([
      [null, null, null, 1, 2, 3, 4],
      [5, 6, 7, 8, 9, 10, 11],
      [12, 13, 14, 15, 16, 17, 18],
      [19, 20, 21, 22, 23, 24, 25],
      [26, 27, 28, 29, null, null, null],
    ]);

    const mar = createWeeksForMonth(2, 2024); // March 2024
    expect(mar).toEqual([
      [null, null, null, null, 1, 2, 3],
      [4, 5, 6, 7, 8, 9, 10],
      [11, 12, 13, 14, 15, 16, 17],
      [18, 19, 20, 21, 22, 23, 24],
      [25, 26, 27, 28, 29, 30, 31],
    ]);

    const apr = createWeeksForMonth(3, 2024); // April 2024
    expect(apr).toEqual([
      [1, 2, 3, 4, 5, 6, 7],
      [8, 9, 10, 11, 12, 13, 14],
      [15, 16, 17, 18, 19, 20, 21],
      [22, 23, 24, 25, 26, 27, 28],
      [29, 30, null, null, null, null, null],
    ]);

    const may = createWeeksForMonth(4, 2024); // May 2024
    expect(may).toEqual([
      [null, null, 1, 2, 3, 4, 5],
      [6, 7, 8, 9, 10, 11, 12],
      [13, 14, 15, 16, 17, 18, 19],
      [20, 21, 22, 23, 24, 25, 26],
      [27, 28, 29, 30, 31, null, null],
    ]);

    const jun = createWeeksForMonth(5, 2024); // June 2024
    expect(jun).toEqual([
      [null, null, null, null, null, 1, 2],
      [3, 4, 5, 6, 7, 8, 9],
      [10, 11, 12, 13, 14, 15, 16],
      [17, 18, 19, 20, 21, 22, 23],
      [24, 25, 26, 27, 28, 29, 30],
    ]);

    const jul = createWeeksForMonth(6, 2024); // July 2024
    expect(jul).toEqual([
      [1, 2, 3, 4, 5, 6, 7],
      [8, 9, 10, 11, 12, 13, 14],
      [15, 16, 17, 18, 19, 20, 21],
      [22, 23, 24, 25, 26, 27, 28],
      [29, 30, 31, null, null, null, null],
    ]);

    const aug = createWeeksForMonth(7, 2024); // August 2024
    expect(aug).toEqual([
      [null, null, null, 1, 2, 3, 4],
      [5, 6, 7, 8, 9, 10, 11],
      [12, 13, 14, 15, 16, 17, 18],
      [19, 20, 21, 22, 23, 24, 25],
      [26, 27, 28, 29, 30, 31, null],
    ]);

    const sep = createWeeksForMonth(8, 2024); // September 2024
    expect(sep).toEqual([
      [null, null, null, null, null, null, 1],
      [2, 3, 4, 5, 6, 7, 8],
      [9, 10, 11, 12, 13, 14, 15],
      [16, 17, 18, 19, 20, 21, 22],
      [23, 24, 25, 26, 27, 28, 29],
      [30, null, null, null, null, null, null],
    ]);

    const oct = createWeeksForMonth(9, 2024); // October 2024
    expect(oct).toEqual([
      [null, 1, 2, 3, 4, 5, 6],
      [7, 8, 9, 10, 11, 12, 13],
      [14, 15, 16, 17, 18, 19, 20],
      [21, 22, 23, 24, 25, 26, 27],
      [28, 29, 30, 31, null, null, null],
    ]);

    const nov = createWeeksForMonth(10, 2024); // November 2024
    expect(nov).toEqual([
      [null, null, null, null, 1, 2, 3],
      [4, 5, 6, 7, 8, 9, 10],
      [11, 12, 13, 14, 15, 16, 17],
      [18, 19, 20, 21, 22, 23, 24],
      [25, 26, 27, 28, 29, 30, null],
    ]);

    const dec = createWeeksForMonth(11, 2024); // December 2024
    expect(dec).toEqual([
      [null, null, null, null, null, null, 1],
      [2, 3, 4, 5, 6, 7, 8],
      [9, 10, 11, 12, 13, 14, 15],
      [16, 17, 18, 19, 20, 21, 22],
      [23, 24, 25, 26, 27, 28, 29],
      [30, 31, null, null, null, null, null],
    ]);
  });
});
